Faster AnnexB conversion
All frames after the first frame only have 1 NALU. So I don't need to do
any memcpy's or searches for the start of the NALUs. Nice :)

I'm waiting on NVIDIA to respond[0] that they only send out single NALUs
after the first frame, but empirically it seems to be the case.

[0] https://devtalk.nvidia.com/default/topic/1069343/gpu-accelerated-libraries/nalu-control-in-nvenc/